+++ UPGRADE.md
@@ -2,4 +2,5 @@
 
 ## v0.1.2
 
-- Moved `WrkFlow\ApiSdkBuilder\ApiFactory` to `WrkFlow\ApiSdkBuilder\Factories\ApiFactory`
+- Moved `WrkFlow\ApiSdkBuilder\ApiFactory` to `WrkFlow\ApiSdkBuilder\Factories\ApiFactory`.
+- Moved `WrkFlow\ApiSdkBuilder\Response` namespace to `WrkFlow\ApiSdkBuilder\Factories\Responses` namespace.
